Fusion style: Set topInset instead of y and height for the background
If the tab button is checked, we want the background to appear raised,
and if not, we want it to appear recessed, with an offset of 2. To do
that, we set y and height in the background delegate of the TabButton,
and binded them to the checked property.
We can do this by setting the topInset instead as it does exactly the
same thing, and reevaluates the height for us, without the extra code.
This is done to not only make the code cleaner, but also as a
workaround to a bug in the evaluation of bindings when setting the
y property in this case.

Fixes: QTBUG-108807
Change-Id: I7b3639deb1162c6bada81e377a36e090dc398654
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 3f6111588e0ab549607ae2eddb9880a87f34f022)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 5732179cd98d0a7b5be16b376bdcab2c8cd801db)
(cherry picked from commit c37f4b0e4273646850a5fb1c3363914e7ae0976c)