Add ContextMenu
This is an attached type that provides a way to open a context menu
in a platform-appropriate manner.

[ChangeLog][Controls] Added ContextMenu. ContextMenu can be attached
to any item in order to show a context menu upon a platform-specific
event, such as a right click or the context menu key.

Fixes: QTBUG-67331
Change-Id: I225a8f498cc5bc18d36c3dc4cf481ef5ef95cd88
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 0794a0e06e40703b191ea3af5ddbb14fafafea8e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>