Add 16-bit/32-bit floating point support for QSGPlainTexture
Part of the HDR story is loading images that contain pixel values in
bigger ranges than between 0 and 1. QQuickDefaultTextureFactory and
QSGPlainTexture automatically converts QImages to RGB(A) 32-bit and
textures to RGB(A/X) 8-bit, which are non-floating point based. This
change adds support for the additional formats required, which are 16
and 32 bit floating point. It will fall back to 16 bit fp textures if
RHI does not support fp 32, and down to the default 8 bits if fp 16 is
not supported either.

Task-number: QTBUG-126035
Change-Id: I6366f04af26126978e087301108af020749bae54
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>