Cleanup pointer grabs when subscene agents are used
When dispatching a pointer event at the end of the method if the event
is a pointer end event we clear up any potential grabbers on the
pointing device.

If a subscene delivery agent exist, or ever existed, we go down a
different path dispatching the initial event there. This path has an
early return which can skip the clearGrabbers line of the main function.

Change-Id: Ib441ac09baa5667fe2e6af64d593f7956a3e2f86
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit ebaf1a1a8a8ffead188fe59e159c123d1a1ff763)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>