Accessibility: Add API to set labelled-by relationship
This is needed when building forms where one text-ish item describes a
control. This allows screenreaders to read the label when focussing the
item. With widgets this can be done using QLabel::setBuddy, but there
is no way to do it in QML

This adds two new properties, labelledBy and labelFor, for both ends
of the relations. Setting either of them automatically fills in the
other direction.

The internal data storage is somewhat generic so that it can be later
extended to also cover other relations like controller, description,
and flows.

Change-Id: Iff0dd387fd03434839038eaeb25967f4b14636a1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>