Abort QQuickPixmap::load if the QQmlEngine is null
If engine is nullptr, calling QQmlEnginePrivate::get(engine)
will cause the program to crash trying to access an
uninitialized object.

The only useful thing this function can do if the engine is null
is to find and return a cached image, if any.

However, in 0290e0fd5285cd48809ade02ce2e67a423ba3152 it was possible
to test QQuickDragAttached without an engine; now there needs to be
one associated via a QQmlContext for this test to continue to pass.

Pick-to: 6.8 6.9
Change-Id: I28596c2063f0474fc97b2e9a3907f8bb38c78cb1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>