QQuickFlickable: don't pause during drag when changing drag direction V2
Flickable/ListView/TableView could sometimes pause (or lag) for almost
a second if you dragged it in one direction, but then switched to
drag it in the opposite direction (without lifting your finger). This
was especially evident on iOS. It could be reproduced in, for
example, the Gallery app (the controls-list in the drawer), or
any other app with a Flickable.

The reason this happened, was because a "reversed" drag would cause
dx and dy to converge back to zero (compared to the starting point
of the drag), which would result in overThreshold becoming false
again. And since the deltas ended up below the threshold, the content
item stopped moving.

The biggest reason for this problem was because we reevaluated the
drag threshold for each subsequent drag in an ongoing flicking
session. This has already been fixed in a previous patch. This patch
will in addition remove the related dx/dy != 0 checks because of the
same reason; Once in dragging mode, the content item should always
move, even if it was dragged exactly back to the starting
point (dx/dy == 0).

Fixes: QTBUG-129599
Change-Id: I0af476903d537269458e2d2ee0d12a90e6dfd9c4
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 3470c058df85faabb1ce0d533c9322d5354d1187)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>