Rearrange tst_qquicktextedit::mouseSelection
This test was always hard to read: it does multiple interactions in
sequence with different sets of property values on TextEdit; and the
single "clicks" argument controlled clicks both before and after the
drag-selection in the middle. At the end, there is also a shift-click.
So we make those steps more explicit. Hopefully it's a little more
maintainable now.

We now wait for the window to be active, in case it helps with
stability.

Pick-to: 6.9 6.8
Task-number: QTBUG-129947
Change-Id: I534c97fc8a3dde87b1b68126c6524fab8fd699ab
Reviewed-by: Frederic Lefebvre <frederic.lefebvre@qt.io>
Reviewed-by: Audun Sutterud <audun.sutterud@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>