CMake: Improve copying speed for QML sources and resources
Previously, each qml source file and resource file of a target was
copied to the target's build dir with a separate cmake custom command
at build time.

For a project with many qml files and images, this can take a long
time, especially for slower generators like Xcode, due to overhead
of spawning a process for each file, and in Xcode's case due to the
additional indirection of wrapper shell script files.

To improve the build time, instead of creating a custom command per
file, create one custom command to copy a target's qml files, and
another one for its resources (like images).

The custom command runs a cmake script that is fed the list of files
to copy. The files are still only copied if they don't exist or are
out of date, preserving dependency information.

For the modified coffee example that is attached to QTBUG-125631 which
contains over 3000 images, on an arm M3 mac, we get the following
speed ups for the initial build:

With ninja, from:
  ninja  23.31s user 206.07s system 1271% cpu 18.046 total
to
  ninja  5.46s user 1.01s system 145% cpu 4.457 total
which is a 4x speed up.

With Xcode, from:
  xcodebuild  27.04s user 1.36s system 42% cpu 1:07.00 total
to
  xcodebuild  0.94s user 0.18s system 7% cpu 14.014 total
and a null build is
  xcodebuild  0.75s user 0.16s system 11% cpu 7.828 total
which is roughly a 5x speed up.

In case of an unforeseen issue, the old way of copying files can still
be enabled by setting the hidden QT_COPY_QML_FILES_OLD_WAY variable to
TRUE.
The intention is to remove the old code path once we can confirm there
are no regressions.

Task-number: QTBUG-125631
Task-number: QTBUG-126205
Task-number: QTBUG-128323
Change-Id: I282919c700d409f0b6eff6a16b734f8502e61611
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit cdd961efc70c659a3e946f5276772ce49a34f644)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>