QQuickTableView: accept rearrange drop event, to avoid misleading animation
As it stood, when you dragged and dropped a column, the
drag image would animate back to the start. This is an
indication that the drop was not accepted, which is not
correct, since the drag'n'drop operation actually
succeeds.

This patch will therefore make sure that we accept
the drop event, to indicate that the operation was
successful. This will stop the drag image from animating
back to start.

Change-Id: I9c39a88520a09b0fd77eb2009f767fdc6be2ffd5
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 257431aed3f238560a785bb7fbf184b14cfb952a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>