dom: add fileLocation entries to comments in FileLocationTree
Add fileLocation entries to comments in the FileLocationTree by
following the current structure of comments in the Dom. This allows
qqmllsutils to actually know sourcelocations of comments and to be able
to retrieve it.

A future commit can make use of that to disable completion in comments,
for example.

Task-number: QTBUG-130694
Change-Id: Ie73e97ca0b99ed54521418ab9757321176664b04
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
(cherry picked from commit 3119dc1f62fc25ae3fb9a1ecc009a76334186dad)
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>