qqmldomastcreator: fix sourcelocations of bindings
Use the sourcelocation of the binding instead of the property definition
when creating bindings. Add a test.

The behavior prior to this commit leads to a bug in a commit in the
relation chain when linking comments to code, as the comment gets linked
to the Binding instead of the property definition.

Change-Id: Id03451967c2093a252c47fe448e5bb46e0358567
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
(cherry picked from commit 40d6c112725a600e4bc019b47b9f8ad7443a1c55)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>