Flickable: when the user is flicking, don't start a drag on a child
While flicking, the user will typically do several subsequent
drags / flicks on the content item to maintain and control the flick.
But as it stood, one of those drags could easily (by accident) hit
a control inside the content item, such as a Slider. And then
he would suddenly be dragging on the Slider instead.

This behavior is unintuitive, and against for example how a web page
in a browser works. Once in a flicking session with an outer
flickable, the user needs to end that session before he can start to
interact with an inner Control.

A flicking session can as such be seen as a single continuous
gesture, even if several drags are practically involved. Once the
Flickable has "won" the first drag, the gesture starts, and no
new gesture/drag negotiation should occur before the gesture ends.

This patch will therefore change the current logic, so that the
user can only start dragging a child item if he's not already
flicking the Flickable.

To do this, this patch will need to handle two different use
cases: child items that handle pointer events, and child items that
have pointer handlers. The former includes e.g QQuickMouseArea and
QQuickControls (and nested flickables), and the latter is
basically items with DragHandlers.

For the item subclass case, the change is done from
For the pointer events case, the change is done from
childMouseEventFilter(): When a flicking session has first
started (isMoving() == true), we filter all pointer events
going to a child, since the child is not the item being
interacted with, and should therefore not receive them.

For the pointer handler case, the solution is basically the
same, except that events going to a pointer handler are not
filtered by childMouseEventFilter(). So the isMoving() test
must also be done from handleMoveEvent(), to ensure that
we don't reevalute the drag threshold for every drag (and open
up for a child DragArea to steal the drag), but instead accept
all drags immediately once a dragging session has started, until
it ends.

Fixes: QTBUG-34881
Change-Id: I0236c9781b908483d4b665e84002aaef32f23e55
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit e116b959452b8d4203077df81e1d1396e0667c60)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>