Add drag and drop to the QtQuick Dialogs SideBar
The current SideBar implementation for the QtQuick Dialogs non-native
FileDialog has a SideBar containing supported standard paths. This patch
adds support for adding favorites to the SideBar, that can be added by
dragging a directory into the SideBar.

Each FileDialogDelegate contains a QQuickTapHandler which listens for a
long press. When long pressed, the delegate can be dragged. A long press
is used to initiate the drag as the delegates are part of a Flickable.
The drag and drop is a system drag and drop. If the dragged delegate is
a directory, the drag is started.

When the dragging enters the SideBar, a delegate appears with 'Add
Favorite' text, to suggest to the user that this is the drop area.

The context menu used for removing a favorite is a QQuickContextMenu,
which handles the context menu event internally.

Change-Id: Icf60f3785522df607ed1f7420486da4318118f0f
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 0b1e5c45de264bf7b62e074c50aba0acb3f157f8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>