Engine: Mark created wrapped objects after GCState::MarkWeakValues
If an object is wrapped after the MarkWeakValues state of the GC but
before the sweep, it will not have been marked and will be collected.

We need to mark wrapped objects when they are created in those cases.

Fixes: QTBUG-130767
Change-Id: I2fd7c8829267a2e3de1ac374859a4d21d948dd8f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit d148d8d78419effdc28b6bf74704379aa9a5950a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>