Mark the focus frame as transparent for positioners
Applies to the focus frame on macOS, Windows and FluentWinUI3 style.

This will ensure that it is excluded by layouts and positioners.
Previously, layouts excluded items with 0x0 size, so setting the size of
the focus frame to 0x0 worked then. This is not the case any longer.

Pick-to: 6.5
Fixes: QTBUG-129427
Change-Id: I399317f8a7a6b941803556e7ffb2208ad3237726
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 7817f81fae52cc893b0f285254a2abbb26c3b16d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>