CMake: Fix copying qml files in in-source builds
When qtdeclarative examples are built in-tree in an in-source build (
as it happens in the 6.8 branch for the TestDocs configuration), none
of the qml files or resource files are copied to the build directory,
because the source and build directories coincide.

In such cases, the ${target}_tooling target is not created,
and the add_dependencies call fails, because the target does not
exist with an error like:
 Cannot add target-level dependencies to non-existent target
 "wearable_tooling".

For now, fix this by just depending on the main target instead.

I will check if it's possible to safely create the tooling target
always, or if there's a better way later.
For now I just want to unblock possible build failures, while also
keeping the P1 fixed.

Amends cdd961efc70c659a3e946f5276772ce49a34f644

Task-number: QTBUG-125631
Task-number: QTBUG-126205
Task-number: QTBUG-128323
Change-Id: I3f641ee5c90d8dd1137631154db3d33bf3e11f38
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit e2bb3c171dc9df4d97e4ebde3452b00047ea3476)