QQuickWidget: resize offscreen window before its content item
Overlays are positioned so that they are centered over the content item
within the window. If the window has a null-size when this is computed,
then the overlay ends up being misplaced, and overlay-provided services,
such as modality or closing popups, stops working for parts of the UI.

Fix this by resizing the offscreen window before resizing the content
item.

Add a second test case that verifies that the overlay geometry is as it
should be, both after the initial show, and after resizing the widget.
The existing resizeOverlay test doesn't use a proper QQuickOverlay, so
it doesn't cover all the relevant code.

Fixes: QTBUG-125135
Pick-to: 6.5
Change-Id: I682d87f07051d213d3bd58c1c2533c84b50a3139
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit f71ef0d8b8d25b80833228fe4e4bfa716bceb398)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>