PopupWindow: Keep better track of the nearest parent item's window
Since Popup::{x,y} are relative to the popup's nearest parent item, any
movement to the parent window (when using popup windows), must cause the
popup to change its x and y properties to reflect this new delta.

Since it's possible to change the popup's parent via JavaScript/QML, and
a new parent item could exist in another window, we'll need to ensure
that we're always connected to the correct window at all times.
Keeping a local reference to this window should be the better approach,
so that we can properly disconnect when the nearest parent item's window
changes, or when we're destroying the popup window.

Users are also occasionally experiencing crashes when using popup
windows. Based on the backtrace, it seems like these crashes are also
related to us not disconnecting correctly when then the popup window's
parent window changes.

Task-number: QTBUG-131898
Pick-to: 6.8
Change-Id: I6b7e3e3d879454f29c0e8665b59f1eca6a88fea2
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 79b354e00614ea6d7740815b94b30713652acdf2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>