qmlls: put import warnings on the correct line
In qmllint, the import warnings are printed right after the "warnings
occurred while importing" warning despite them having no sourcelocation.

This warning gets a default sourcelocation when used in qmlls which
points to the first line of the file, so the "Failed to import" warning
ends up an the first line of a file, and not on the same line as the
"warnings occurred while importing" warning. This is confusing for users.

Modify a test to expect that the "Failed to import" warning has a valid
sourcelocation, so that it is shown on the correct line in qmlls.

Fixes: QTBUG-131958
Pick-to: 6.8
Change-Id: Ia0c327b82e89cd0a003d05b4c3efe693fb3b3ed7
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 7188ee3513f1ba2d9f745793b6db3c8db374191e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>