Re-use safe area if we've already created one for the target item
The qmlAttachedProperties cache is based on the object the attached
property is used on, but we attach the safe area to the associated
Item. We want the request for both Window and Window.contentItem
to result in the same safe area, so do a check for that explicitly.

Change-Id: I17d936f2b7c6aa0d4cc46b800e097d1472a5a784
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 576e5fce84b0e554e16949e9e825144438b5acc8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>