Update palette change to the popup content item placed within item view
The popup content item has not been updated with the changed palette in
all the cases. For instance, when the popup has been configured as part
of the list view, it has not been considered as the child item of the
window but the content item and in those cases the change in the  
palette has not been updated to the item configured within the popup.  

This is because we update the palette of the popup content item only if
its child of the window. For other cases, we still set the palette to
the popup through (updateChildrenPalette), but the popup doesn't
propagate this to the content item.

This patch ensures that the popup updates its palette to its content
item whenever its palette has been changed.

Fixes: QTBUG-128462
Pick-to: 6.7 6.5
Change-Id: I4eacb8c8e2b1ae95bb938bc2e72b8a4f48125023
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit f6648c67c1018d0f2331eeeabfeccddd6f88c704)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit e9cb21dac82591bf9e2a3de781898095bfdd1688)