QQuickWindowContainer: Give focus to embedded window on FocusIn event
When the QQuickWindowContainer gains focus from a tab key event, the
focus would remain within the QQuickWindowContainer item and it would
be impossible to tab focus from the container item to the embedded
window's children or from the parent window to embedded window.
This happened because the window and its content item are not
children of the window container item, hence not its prev/next in the
tab focus chain.
To fix, give focus to the embedded window on a focusIn event by
calling requestActivate() and make sure to determine the right focus
item to give focus to in case the focus reason was from a tab/backtab
key event.

Fixes: QTBUG-132071
Change-Id: I5ea9cacaca7dce85285817978396e7297bc54baf
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit c815f6fe7b78f38bde14da1300bb791c5eaae8d7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>