Remove a lie from the documentation about property initialization order
In most cases, the behavior that was described is not the observed
behavior anymore (today, the first MyWidget would normally get the
focus). This has been the case since at least 5.15

This is because QML doesn't guarantee which element order it will
initialize properties.  It has been like this for a long time. Maybe the
documentation was written before that.

Pick-to: 6.8 6.5
Change-Id: Ibe4ead198e39e16ac2cfb3b841f4f0697172690a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 953c38b5818231d52ea170eb6466296f94473c76)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>