FluentWinUI3: Show focus frame when focus reason is from arrow keys
We don't have a Qt::FocusReason enumerator value that represents the
focus reason being arrow keys, so for cases that an item gains focus
due to arrow-key navigating, the focus reason will be OtherFocusReason.
The FluentWinUI3 style only draws the focus frame when the focus
reason was Tab/Backtab to avoid drawing it for focus reasons other
than keyboard focus, but it should draw it for when focus shifts due
to arrow keys too, so include Qt::OtherFocusReasons in the conditions.
Also, reset the focus frame when we can't find a focus object AND
when the focus reason is none of the above in order to avoid showing
the frame around an item that lost focus because of another reason.

Fixes: QTBUG-130328
Pick-to: 6.8
Change-Id: I6cbc13124a0e672bf453896bafd0ba1d8656449e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit d89045996da43c55cf9bbe80563b6a02d307cf95)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>